pyproject.toml
[tool.poetry]
name = "django-forum"
version = "0.1.0"
description = ""
authors = ["John-Hugh Hedrick <johnhugh1998@gmail.com>", "Matthew Purnell <mattpurnell98@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.*"
django = "1.11.25"
django-recaptcha = "^2.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
